Improve word rendering
Previously we'd use a word's Y-Postion and its height as the font-size
in pixels. This led to a very irregular line appearance, since words
consisting of all-lowercase letter would be rendered far too small.
With this change, we now use the line's Y-position and height instead,
making for a far more regular line appearance, with no problems for
selecting text.