Improve handling of spaces, test selection
Previously our approach to parsing spaces would differ between ALTO and
hOCR. In the former, we'd have empty word objects, in the latter the
whitespace would be included in the text of its previous word. This has
now been unified to always be part of the previous word. Also included
now is a test for the selection functionality.