Render each word as its own <text> element (fix #1, #5)
Testing revealed that wrapping each line in a <text> and rendering
individual words as `<tspan>` elements did not have any effect on the
text selection behavior. Due to an incomplete SVG implementation in
Firefox, it would not stretch `<tspan>` elements to the appropriate
width.

So, with no ill effects to expect (to my current knowledge), this commit
removes the rendering of `<tspan>` elements and renders every word as an
individual, unwrapped `<text>` element.