Unify space handling between ALTO and hOCR.
Previously, we would add word tokens for space characters if an ALTO
document encoded spaces, and otherwise expand existing words with
whitespace if it didn't (always the case for hOCR).

This required a bunch of guesstimating and brittle hacks, so now we
instead opt to use the space between existing words for whitespace
uniformly.

This also renames `word` (which would now be misleading, since it can
also contain only whitespace) to `span`.