Always swallow pointerdown events if selection is enabled.
Previously this was disabled for mobile devices, since we wanted for the
user to be able to pan and zoom the OSD canvas despite selection being
enabled. It turns out that this was very wonky across browsers, so we've
enabled event-swallowing again.
This has the effect that, if selection is enabled, panning and zooming
has to be initiated on parts of the canvas that are not covered by text.